Refactor auth routing into modular package
